+++ tests/eta_tests.py
@@ -1,5 +1,5 @@
 from steerclear.eta import *
-from steerclear.models import Ride
+from steerclear.api.models import Ride
 import unittest, json, urllib, vcr, requests, os
 
 # vcr object used to record api request responses or return already recorded responses
